Move some reward handling to the API module
