+++ even-more-fish-api/src/main/java/com/oheers/fish/api/reward/EMFRewardsLoadEvent.java
@@ -1,0 +1,23 @@
+package com.oheers.fish.api.reward;
+
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+import org.jetbrains.annotations.NotNull;
+
+/**
+ * Designed for plugins to re-register their reward types if the plugin is reloaded with (for example) PlugManX
+ */
+public class EMFRewardsLoadEvent extends Event {
+
+    private static HandlerList handlers = new HandlerList();
+
+    public static HandlerList getHandlerList() {
+        return handlers;
+    }
+
+    @Override
+    public @NotNull HandlerList getHandlers() {
+        return handlers;
+    }
+
+}
