+++ even-more-fish-api/src/main/java/com/oheers/fish/api/reward/Reward.java
@@ -1,7 +1,5 @@
-package com.oheers.fish.competition.reward;
+package com.oheers.fish.api.reward;
 
-import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.api.EMFRewardEvent;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
@@ -19,7 +17,7 @@ public class Reward {
     public Reward(@NotNull String identifier) {
         String[] split = identifier.split(":");
         if (split.length < 2) {
-            EvenMoreFish.getInstance().getLogger().warning(value + " is not formatted correctly. It won't be given as a reward");
+            RewardManager.getInstance().getLogger().warning(value + " is not formatted correctly. It won't be given as a reward");
             this.key = "";
             this.value = "";
         } else {
