+++ even-more-fish-api/src/main/java/com/oheers/fish/api/reward/RewardType.java
@@ -1,4 +1,4 @@
-package com.oheers.fish.competition.reward;
+package com.oheers.fish.api.reward;
 
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
