+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/rewardtypes/EXPRewardType.java
@@ -1,28 +1,34 @@
-package com.oheers.fish.competition.reward.types;
+package com.oheers.fish.competition.rewardtypes;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.competition.reward.RewardType;
-import com.oheers.fish.config.messages.OldMessage;
+import com.oheers.fish.api.reward.RewardType;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 
-public class MessageRewardType implements RewardType {
+public class EXPRewardType implements RewardType {
 
     @Override
     public void doReward(@NotNull Player player, @NotNull String key, @NotNull String value, Location hookLocation) {
-        player.sendMessage(new OldMessage().setMSG(value).setReceiver(player).toString());
+        int experience;
+        try {
+            experience = Integer.parseInt(value);
+        } catch (NumberFormatException ex) {
+            EvenMoreFish.getInstance().getLogger().warning("Invalid number specified for RewardType " + getIdentifier() + ": " + value);
+            return;
+        }
+        player.giveExp(experience);
     }
 
     @Override
     public @NotNull String getIdentifier() {
-        return "MESSAGE";
+        return "EXP";
     }
 
     @Override
     public @NotNull String getAuthor() {
-        return "Oheers";
+        return "FireML";
     }
 
     @Override
