+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/rewardtypes/ItemRewardType.java
@@ -1,8 +1,8 @@
-package com.oheers.fish.competition.reward.types;
+package com.oheers.fish.competition.rewardtypes;
 
 import com.oheers.fish.EvenMoreFish;
 import com.oheers.fish.FishUtils;
-import com.oheers.fish.competition.reward.RewardType;
+import com.oheers.fish.api.reward.RewardType;
 import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
