+++ even-more-fish-plugin/src/main/java/com/oheers/fish/competition/rewardtypes/PermissionRewardType.java
@@ -1,7 +1,7 @@
-package com.oheers.fish.competition.reward.types;
+package com.oheers.fish.competition.rewardtypes;
 
 import com.oheers.fish.EvenMoreFish;
-import com.oheers.fish.competition.reward.RewardType;
+import com.oheers.fish.api.reward.RewardType;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
